# frontend/openapi.json
{
  "components": {
...
    "securitySchemes": {
      "bearer": {
        "bearerFormat": "JWT",
        "description": "A JWT signed with the secret of the server, sent as `Authorization: Bearer <token>` (`Authorization: Token <token>`, which InfluxDB clients send, works too). Each operation lists the scope it needs: `read`, `write`, `delete` or `admin`. Make one with `sensapp generate-token`, or with `POST /api/v1/admin/tokens` and an `admin` token (the Credentials tab of the web UI does). With authentication disabled (`SENSAPP_AUTH_DISABLED`) no token is needed.",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "Sensor data ingestion, storage, and retrieval platform",
    "license": {
      "identifier": "Apache-2.0",
      "name": "Apache-2.0"
    },
    "title": "sensapp",
    "version": "0.3.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/series": {
      "get": {
        "operationId": "list_series",
        "parameters": [
          {
            "description": "Filter series by metric name",
            "in": "query",
            "name": "metric",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "PromQL-style label selector (e.g., '{env=\"prod\",region=~\"us.*\"}')",
            "in": "query",
            "name": "selector",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Series per page, 256 by default, 16384 at most",
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "minimum": 0,
              "type": "integer"
            }
          },
          {
            "description": "Cursor of the next page: the `bookmark` of the `hydra:next` link of the previous page",
            "in": "query",
            "name": "bookmark",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Time series catalog in DCAT format. A full page has a `hydra:view` with the link to the next page"
          },
          "503": {
            "description": "The storage backend is unavailable: retry later"
          },
          "504": {
            "description": "The request took longer than SENSAPP_HTTP_SERVER_TIMEOUT_SECONDS, usually because the storage backend hangs"
          }
        },
        "security": [
          {
            "bearer": [
              "read"
            ]
          }
        ],
        "summary": "List all series (time series) in DCAT catalog format.",
        "tags": [
          "SensApp"
        ]
      }
    },
...
  },
...
}